utils: change return value of parser_read_file_line
Return 1 if some data was read, 0 if end of file reached
and no data was read but no error occured, and -1 if an
error occured.  The previous behaviour was to return the
number of bytes read, or -1 in case of error, which didn't
made any difference between reading a blank line and reaching
the end of the file.  Now a blank line returns 1 as a line
feed character has been read from the file.