main.c: use new pl_gpio API + tidy-up"
Initialise the new pl_gpio API using msp430-gpio and pass it
to other modules that need to use GPIOs.  Some store is as a
static variable to reduce the amount of refactoring.

Note: There is a problem with the VCOM + HW info EEPROM on
Hummingbird Z6-Z7, this fails to read until the Epson has
been reset (probably causing an I2C master conflict).